Extract usePrivilegedUsersTableData hook from table component
