+++ x-pack/solutions/security/plugins/security_solution/public/entity_analytics/components/privileged_user_monitoring/components/privileged_users_table/index.tsx
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-import React, { useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import {
   useEuiTheme,
   EuiPanel,
@@ -19,29 +19,17 @@ import {
   EuiCallOut,
 } from '@elastic/eui';
 import { css } from '@emotion/react';
-import { take } from 'lodash/fp';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
-import { useQuery } from '@tanstack/react-query';
-import { getESQLResults, prettifyQuery } from '@kbn/esql-utils';
 import { i18n } from '@kbn/i18n';
-import { getPrivilegedMonitorUsersIndex } from '../../../../../../common/entity_analytics/privilege_monitoring/utils';
 import { useGlobalTime } from '../../../../../common/containers/use_global_time';
 import { useQueryInspector } from '../../../../../common/components/page/manage_query';
 import { useQueryToggle } from '../../../../../common/containers/query_toggle';
-import { useRiskScore } from '../../../../api/hooks/use_risk_score';
-import type { TableItemType } from './types';
-import { getPrivilegedUsersQuery } from './esql_source_query';
 import { UserPanelKey } from '../../../../../flyout/entity_details/shared/constants';
-import type { EntityRiskScore } from '../../../../../../common/search_strategy';
-import { buildEntityNameFilter, EntityType } from '../../../../../../common/search_strategy';
+
 import { buildPrivilegedUsersTableColumns } from './columns';
-import { esqlResponseToRecords } from '../../../../../common/utils/esql';
-import { useKibana } from '../../../../../common/lib/kibana';
-import { useGlobalFilterQuery } from '../../../../../common/hooks/use_global_filter_query';
 import { HeaderSection } from '../../../../../common/components/header_section';
-import { useAssetCriticalityFetchList } from '../../../asset_criticality/use_asset_criticality';
-import type { CriticalityLevelWithUnassigned } from '../../../../../../common/entity_analytics/asset_criticality/types';
+import { usePrivilegedUsersTableData } from './hooks';
 
 export const DEFAULT_PAGE_SIZE = 10;
 
@@ -71,152 +59,18 @@ const useOpenUserFlyout = () => {
   };
 };
 
-interface RiskScoresByUserName {
-  [key: string]: EntityRiskScore<EntityType.user>;
-}
-
-interface AssetCriticalityByUserName {
-  [key: string]: CriticalityLevelWithUnassigned;
-}
-
 export const PrivilegedUsersTable: React.FC<{ spaceId: string }> = ({ spaceId }) => {
   const { setQuery, deleteQuery } = useGlobalTime();
   const { toggleStatus, setToggleStatus } = useQueryToggle(PRIVILEGED_USERS_TABLE_QUERY_ID);
 
   const { euiTheme } = useEuiTheme();
 
-  const { data } = useKibana().services;
-
   const openUserFlyout = useOpenUserFlyout();
 
   const columns = buildPrivilegedUsersTableColumns(openUserFlyout, euiTheme);
   const [currentPage, setCurrentPage] = useState<number>(1);
-
-  const privilegedUsersTableQuery = getPrivilegedUsersQuery(
-    spaceId,
-    currentPage * DEFAULT_PAGE_SIZE + 1 // we add 1 so that we know if there are more results to show
-  );
-
-  const { filterQuery: filterQueryWithoutTimerange } = useGlobalFilterQuery();
-
-  const {
-    data: result,
-    isLoading: loadingPrivilegedUsers,
-    isError: privilegedUsersError,
-    refetch,
-  } = useQuery({
-    queryKey: ['privileged-users-table', privilegedUsersTableQuery, filterQueryWithoutTimerange],
-    enabled: toggleStatus,
-    queryFn: async () => {
-      return getESQLResults({
-        esqlQuery: privilegedUsersTableQuery,
-        search: data.search.search,
-        filter: filterQueryWithoutTimerange,
-      });
-    },
-  });
-
-  const records = useMemo(() => esqlResponseToRecords<TableItemType>(result?.response), [result]);
-
-  const nameFilterQuery = useMemo(() => {
-    const userNames = records.map((user) => user['user.name']);
-    return buildEntityNameFilter(EntityType.user, userNames);
-  }, [records]);
-
-  const {
-    data: riskScoreData,
-    error: riskScoreError,
-    loading: loadingRiskScore,
-    hasEngineBeenInstalled: hasRiskEngineBeenInstalled,
-  } = useRiskScore<EntityType.user>({
-    riskEntity: EntityType.user,
-    filterQuery: nameFilterQuery,
-    onlyLatest: true,
-    pagination: {
-      cursorStart: 0,
-      querySize: records.length,
-    },
-    skip: nameFilterQuery === undefined || records.length === 0,
-  });
-
-  const riskScores = riskScoreData && riskScoreData.length > 0 ? riskScoreData : [];
-
-  const riskScoreByUserName: RiskScoresByUserName = Object.fromEntries(
-    riskScores.map((riskScore) => [riskScore.user.name, riskScore])
-  );
-  const {
-    data: assetCriticalityData,
-    isError: assetCriticalityError,
-    isLoading: loadingAssetCriticality,
-  } = useAssetCriticalityFetchList({
-    idField: 'user.name',
-    idValues: records.map((user) => user['user.name']),
-    skip: !toggleStatus || records.length === 0,
-  });
-  const assetCriticalityRecords =
-    assetCriticalityData && assetCriticalityData.records.length > 0
-      ? assetCriticalityData.records
-      : [];
-
-  const assetCriticalityByUserName: AssetCriticalityByUserName = Object.fromEntries(
-    assetCriticalityRecords.map((assetCriticalityRecord) => [
-      assetCriticalityRecord.id_value,
-      assetCriticalityRecord.criticality_level,
-    ])
-  );
-
-  const enrichedRecords: TableItemType[] = useMemo(
-    () =>
-      records.map((record, index) => {
-        let enrichedFields = {};
-
-        const riskScore: EntityRiskScore<EntityType.user> | undefined =
-          riskScoreByUserName[record['user.name']];
-        if (riskScore) {
-          enrichedFields = {
-            ...enrichedFields,
-            risk_score: riskScore.user.risk.calculated_score_norm,
-            risk_level: riskScore.user.risk.calculated_level,
-          };
-        }
-
-        const assetCriticality: CriticalityLevelWithUnassigned | undefined =
-          assetCriticalityByUserName[record['user.name']];
-
-        if (assetCriticality) {
-          enrichedFields = {
-            ...enrichedFields,
-            criticality_level: assetCriticality,
-          };
-        }
-
-        return {
-          ...record,
-          ...enrichedFields,
-        };
-      }),
-    [records, riskScoreByUserName, assetCriticalityByUserName]
-  );
-
-  const isLoading =
-    loadingPrivilegedUsers || (records.length > 0 && (loadingRiskScore || loadingAssetCriticality));
-  const visibleRecords = take(currentPage * DEFAULT_PAGE_SIZE, enrichedRecords);
-
-  const inspect = useMemo(() => {
-    return {
-      dsl: [
-        JSON.stringify(
-          {
-            index: [getPrivilegedMonitorUsersIndex(spaceId)],
-            body: prettifyQuery(privilegedUsersTableQuery, false),
-          },
-          null,
-          2
-        ),
-      ],
-      response: result ? [JSON.stringify(result, null, 2)] : [],
-    };
-  }, [privilegedUsersTableQuery, result, spaceId]);
+  const { visibleRecords, isLoading, hasError, refetch, inspect, hasNextPage } =
+    usePrivilegedUsersTableData(spaceId, currentPage, toggleStatus);
 
   useQueryInspector({
     deleteQuery,
@@ -224,8 +78,9 @@ export const PrivilegedUsersTable: React.FC<{ spaceId: string }> = ({ spaceId })
     refetch,
     setQuery,
     queryId: PRIVILEGED_USERS_TABLE_QUERY_ID,
-    loading: loadingPrivilegedUsers,
+    loading: isLoading,
   });
+
   return (
     <EuiPanel hasBorder hasShadow={false} data-test-subj="privileged-users-table-panel">
       <HeaderSection
@@ -238,9 +93,7 @@ export const PrivilegedUsersTable: React.FC<{ spaceId: string }> = ({ spaceId })
         outerDirection="column"
         hideSubtitle
       />
-      {(privilegedUsersError ||
-        (hasRiskEngineBeenInstalled && riskScoreError) ||
-        assetCriticalityError) && (
+      {hasError && (
         <EuiCallOut
           title={i18n.translate(
             'xpack.securitySolution.entityAnalytics.privilegedUserMonitoring.privilegedUsersTable.error',
@@ -255,13 +108,8 @@ export const PrivilegedUsersTable: React.FC<{ spaceId: string }> = ({ spaceId })
       )}
       {toggleStatus && (
         <EuiFlexGroup direction="column" gutterSize="s">
-          <EuiFlexItem>
-            {isLoading &&
-              (loadingPrivilegedUsers || loadingRiskScore || loadingAssetCriticality) && (
-                <EuiProgress size="xs" color="accent" />
-              )}
-          </EuiFlexItem>
-          {records.length > 0 ? (
+          <EuiFlexItem>{isLoading && <EuiProgress size="xs" color="accent" />}</EuiFlexItem>
+          {visibleRecords.length > 0 ? (
             <>
               <EuiText size={'s'}>
                 <FormattedMessage
@@ -299,7 +147,7 @@ export const PrivilegedUsersTable: React.FC<{ spaceId: string }> = ({ spaceId })
               </EuiText>
             )
           )}
-          {records.length > currentPage * DEFAULT_PAGE_SIZE && (
+          {hasNextPage && (
             <EuiFlexItem grow={false}>
               <EuiButtonEmpty
                 isLoading={isLoading}
